Moving DB and SensingEngine singletons to their respective classes
